Added pie-chart and polygon shapes.
Now nodes and sprites can have a "pie-chart" shape. The pie-chart
requires to store on the corresponding element a "ui.pie-values" attribute
that definesi real values between 0 and 1 whose sum must be 1. The
pie-chart colors will be taken from the "fill-color" CSS property:

node {
	shape: pie-chart;
	fill-mode: plain;
	fill-color: red, green, blue;
}

In the program, for example:

Node A = graph.getNode("A");
A.addAttribute("ui.pie-values", 0.1f, 0.3f, 0.6f);

One can also now choose an arbitrary shape for nodes and sprites
(edges are in the works) using the "polygon" shape. As for pie-charts,
polygons use a special attribute stored on the node or on the edge
nammed "ui.points". This must be a sequence of 3D points either under
the form of real numbers (any form transformable to a Number is accepted)
by pack of three, or a sequence of Point3 (in org.graphstream.ui.geom).

node {
	shape: polygon;
}

in the program, for example:

A.addAttribute("ui.points",
	new Point3(-1, -1, 0),
	new Point3( 1, -1, 0),
	new Point3( 0,  1, 0));

Will render the node as a triangle.