Base work for shaped edges.
The basis for shapes described by poly-lines is here. There are two
kinds of polylines. The first (shape: polyline) requires the points of
the edge to be at the absolute position in space. This may not be very
useful for dynamic graphs, however should be really adapted for GIS
(geographic) data for example. This first mode is functionnal.

The second (shape: polyline-scaled) sees the "ui.points" as a vector
shape that can be rotated along the edge vector, defined by its two
endpoints at its nodes position, and is scaled to fit between the two
endpoints. This mode is not yet enabled and needs more work.