Corrected a bug in text rendering.
The text bounds computation is cached in order to avoid recomputing
it at each new redraw, using the TextLayout interface of Java. However
this text layout uses a FontRenderContext dependant on the current
affine transform of the Graphics2D. This one was not the good one
(we use an identity transform to render text).