Basis for an OpenGL GraphRenderer and View named HobGobelin.
Added the hobgobelin package.
Added a new JoglView that allows to open a GLCanvas and provide a GL
context.
Added a new HobGobelinGraphRenderer that is compatible with the
JoglView and can render using the GL context of the view.

There are bases to build a camera, and various renderers for nodes,
edges, sprites, etc.